requirements.txt
#core ML
torch>=2.0.0
#scikit-learn>=1.3.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
evaluate>=0.4.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
torchaudio>=2.0.0
audiomentations
torchcodec

# data science

numpy>=1.24.0
pandas>=2.0.0

# logging / visualization
matplotlib>=3.7.0
seaborn
# utilities
pyyaml>=6.0
python-dotenv>=1.0.0

